requirements.txt
flask==2.0.2
pandas
python-dotenv
gunicorn>=20.1.0
